Rename NavigationManager.
